+++ rollem-discord.code-workspace
@@ -4,5 +4,9 @@
 			"path": "."
 		}
 	],
-	"settings": {}
+	"settings": {
+		"files.exclude": {
+			"node_modules": true
+		}
+	}
 }
