+++ src/rollem-bot/behaviors/ping-pong.behavior.ts
@@ -1,9 +1,18 @@
 import { BehaviorBase } from "./behavior-base";
+import { injectable } from "inversify";
+import { Client } from "discord.js";
+import { Logger } from "@bot/logger";
 
 /**
  * Replies with "pong" whem DM'd "ping".
  */
+@injectable()
 export class PingPongBehavior extends BehaviorBase {
+  constructor(
+    protected readonly client: Client,
+    protected readonly logger: Logger,
+  ) { super(client, logger); }
+
   protected register() {
     this.client.on('message', message => {
       if (message.author.bot) { return; }
